Add DeviceAuthManager that has the common code. Platform specific classes do only getting the cert and signing.
